feat(theme-persistence): implementa persistência do tema
